+++ index.ts
@@ -1,8 +1,8 @@
 import { type BunPlugin } from 'bun'
-import { resolveCallback } from './utils/utis'
+import { resolveCallback } from './utils/utils'
 
 export default (mappers: { [key: string]: string }): BunPlugin => ({
-	name: 'BunModuelNameMapperPlugin',
+	name: 'BunModuleNameMapperPlugin',
 	setup(build) {
 		const filter = new RegExp(Object.keys(mappers).join('|'))
 		build.onResolve({ filter }, resolveCallback(mappers))
